+++ app/src/androidTest/java/woowacourse/payments/CardNumberTest.kt
@@ -8,6 +8,7 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import woowacourse.payments.ui.DigitTextField
 
 class CardNumberTest {
     @get:Rule
