+++ app/src/androidTest/java/woowacourse/payments/CardOwnerTest.kt
@@ -7,6 +7,7 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import woowacourse.payments.ui.LimitedTextField
 
 class CardOwnerTest {
     @get:Rule
