+++ app/src/main/java/woowacourse/payments/list/ListActivity.kt
@@ -40,7 +40,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import woowacourse.payments.newCard.NewCardActivity
-import woowacourse.payments.PaymentCard
+import woowacourse.payments.ui.PaymentCard
 import woowacourse.payments.domain.Card
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
